fadebackground added to image, resolved conflicts
