+++ src/morphocut/pipeline/__init__.py
@@ -2,5 +2,4 @@
 Processing nodes are generators.
 """
 
-from morphocut.pipeline.annotation import FadeBackground, DrawContours
 from morphocut.pipeline.object_scale import ObjectScale
